auth_service/pyproject.toml
[project]
name = "django-auth-service"
version = "0.1.0"
description = "Django authentication microservice"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "mongoengine (>=0.29.1,<0.30.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]

[tool.poetry]
packages = [
    {include = "src"}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
